Updated scripts to pass tests.
By default, the script will not crash when an atom can not be
found but rather print a warning. This allows to produce most
of the output, also in case an atom is not present (e.g. for
protonated His residues).

Updated the reference data that in most case becomes longer
since the script now continues to work.